feat: introduce ChunkDHTValue and update related models for chunk handling
