+++ storb/dht/piece_dht.py
@@ -2,27 +2,14 @@ from pydantic import BaseModel
 
 from storb.util.piece import PieceType
 
-# infohash -> ChunkDHTValue -> PieceDHTValue
-
 
 class PieceDHTValue(BaseModel):
+    piece_hash: str
     miner_id: int
     chunk_idx: int
     piece_idx: int
     piece_type: PieceType
-
-    def to_dict(self) -> dict:
-        return self.model_dump()
-
-
-class ChunkDHTValue(BaseModel):
-    piece_hashes: list[str]
-    chunk_idx: int
-    k: int
-    m: int
-    chunk_size: int
-    padlen: int
-    original_chunk_length: int
+    signature: str
 
     def to_dict(self) -> dict:
         return self.model_dump()
